NMR-5167: allow opening multiple FIDs in a grid, moved processing widgets from FXMLController to ProcessorController, removed processorcontroller from being an object of fxmlcontroller and the reverse. Switching between charts on a grid updates spectrum status bar, changed System.outs to logs
