--- nmrfx-processor-gui/NMRScene.fxml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import java.net.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import org.controlsfx.control.*?>
-<?import org.nmrfx.processor.gui.*?>
-
-<BorderPane fx:id="mainBox" prefHeight="600.0" prefWidth="900.0" styleClass="mainFxmlClass" xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.nmrfx.processor.gui.FXMLController">
-    <stylesheets>
-        <URL value="@/styles/Styles.css" />
-    </stylesheets>
-    <top>
-        <VBox fx:id="topBar" GridPane.hgrow="ALWAYS">
-            <children>
-                <ToolBar fx:id="toolBar" prefWidth="200.0" BorderPane.alignment="CENTER">
-                    <items>
-                    </items>
-                </ToolBar>
-            </children>
-        </VBox>
-    </top>
-    <center>
-        <BorderPane fx:id="borderPane" HBox.hgrow="ALWAYS">
-            <top>
-            </top>
-            <center>
-                <StackPane fx:id="chartPane" VBox.vgrow="ALWAYS" prefHeight="1000" />
-            </center>
-            <left>
-                <VBox fx:id="vectorBox" BorderPane.alignment="CENTER">
-                    <children>
-                        <Slider fx:id="vecNum1" blockIncrement="1.0" majorTickUnit="16.0" min="1" max="64.0" minorTickCount="4" onValueChange="#handleVecNum" onMouseReleased="#handleVecRelease" orientation="VERTICAL" showTickLabels="true" showTickMarks="true" VBox.vgrow="ALWAYS" />
-                    </children>
-                </VBox>
-            </left>
-            <right>
-                <GridPane fx:id="rightBox" BorderPane.alignment="CENTER">
-                    <columnConstraints>
-                        <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" />
-                        <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" />
-                    </columnConstraints>
-                    <rowConstraints>
-                        <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-                    </rowConstraints>
-                    <children>
-                        <VBox fx:id="phaserBox" GridPane.columnIndex="0" fillWidth="false" GridPane.hgrow="NEVER" GridPane.vgrow="ALWAYS">
-                        </VBox>
-                    </children>
-                </GridPane>
-            </right>
-            <bottom>
-                <HBox BorderPane.alignment="CENTER">
-                    <children>
-                        <Label fx:id="rowLabel" />
-                    </children>
-                </HBox>
-            </bottom>
-        </BorderPane>
-    </center>
-    <right>
-    </right>
-    <bottom>
-        <VBox fx:id="bottomBox">
-            <children>
-                <ToolBar fx:id="btoolBar" >
-                    <items>
-                        <Pane HBox.hgrow="ALWAYS" />
-                    </items>
-                </ToolBar>
-            </children>
-        </VBox>
-    </bottom>
-</BorderPane>
